+++ backend/substrapp/views/computetask.py
@@ -25,7 +25,6 @@ from localrep.serializers import ComputeTaskSerializer as ComputeTaskRepSerializ
 from localrep.serializers import ComputeTaskWithRelationshipsSerializer as ComputeTaskWithRelationshipsRepSerializer
 from substrapp import exceptions
 from substrapp.orchestrator import get_orchestrator_client
-from substrapp.utils import get_owner
 from substrapp.views.filters_utils import CharInFilter
 from substrapp.views.filters_utils import ChoiceInFilter
 from substrapp.views.filters_utils import MatchFilter
@@ -35,9 +34,6 @@ from substrapp.views.utils import TASK_CATEGORY
 from substrapp.views.utils import ApiResponse
 from substrapp.views.utils import ValidationExceptionError
 from substrapp.views.utils import get_channel_name
-from substrapp.views.utils import permissions_intersect
-from substrapp.views.utils import permissions_union
-from substrapp.views.utils import to_string_uuid
 from substrapp.views.utils import validate_key
 
 logger = structlog.get_logger(__name__)
@@ -78,97 +74,6 @@ def _compute_extra_data(orc_task, task_data):
 
     else:
         raise Exception(f'Task type "{orc_task["category"]}" not handled')
-
-
-def _get_task_outputs(channel, task_data, trunk_permissions, tasks_cache=None):
-    """
-    This function replicates the historical permission rules, originally implemented in the orchestrator.
-    In a later iteration, this function will be removed and the permissions will be set explicitly by the client.
-    """
-    if task_data["category"] == ComputeTaskRep.Category.TASK_TRAIN:
-        # Adapted from
-        # https://github.com/owkin/orchestrator/blob/a9f7d14867746f58958f30dd6304167d87a6f15c/lib/service/computetask.go#L506
-        with get_orchestrator_client(channel) as client:
-            algo = client.query_algo(task_data["algo_key"])
-            data_manager = client.query_datamanager(task_data["train"]["data_manager_key"])
-
-        permissions = permissions_intersect(algo["permissions"]["process"], data_manager["permissions"]["process"])
-        return {"model": {"permissions": permissions}}
-
-    elif task_data["category"] == ComputeTaskRep.Category.TASK_PREDICT:
-        # Only give permission to worker
-        with get_orchestrator_client(channel) as client:
-            data_manager = client.query_datamanager(task_data["predict"]["data_manager_key"])
-        return {"predictions": {"permissions": {"public": False, "authorized_ids": [data_manager["owner"]]}}}
-
-    elif task_data["category"] == ComputeTaskRep.Category.TASK_TEST:
-        # Performances should always be public
-        # Orchestrator will refuse anything else
-        return {"performance": {"permissions": {"public": True}}}
-
-    elif task_data["category"] == ComputeTaskRep.Category.TASK_AGGREGATE:
-        # Adapted from
-        # https://github.com/owkin/orchestrator/blob/a9f7d14867746f58958f30dd6304167d87a6f15c/lib/service/computetask.go#L456-L471
-        permissions = {
-            "public": False,
-            "authorized_ids": [get_owner()],
-        }
-        for task_key in task_data["parent_task_keys"]:
-            perm = _get_parent_task_output_permission(channel, task_key, tasks_cache)
-            permissions = permissions_union(permissions, perm)
-
-        return {"model": {"permissions": permissions}}
-
-    elif task_data["category"] == ComputeTaskRep.Category.TASK_COMPOSITE:
-        # Adapted from
-        # https://github.com/owkin/orchestrator/blob/a9f7d14867746f58958f30dd6304167d87a6f15c/lib/service/computetask.go#L409-L416
-        with get_orchestrator_client(channel) as client:
-            data_manager = client.query_datamanager(task_data["composite"]["data_manager_key"])
-
-        permissions_local = {"public": False, "authorized_ids": [data_manager["owner"]]}
-        permissions_shared = trunk_permissions
-
-        if not permissions_shared["public"]:
-            authorized_ids = set(permissions_shared["authorized_ids"])
-            authorized_ids.add(data_manager["owner"])
-            permissions_shared["authorized_ids"] = list(authorized_ids)
-
-        return {
-            "local": {"permissions": permissions_local},
-            "shared": {"permissions": permissions_shared},
-        }
-
-    else:
-        raise Exception(f'Task type "{task_data["category"]}" unknown')
-
-
-def _get_parent_task_output_permission(channel, task_key, tasks_cache):
-    task = None
-    perm = None
-    from_orchestrator = False
-
-    if tasks_cache:
-        task = tasks_cache.get(to_string_uuid(task_key))
-
-    if not task:
-        from_orchestrator = True
-        with get_orchestrator_client(channel) as client:
-            task = client.query_task(task_key)
-
-    if task["category"] == ComputeTaskRep.Category.TASK_TRAIN:
-        perm = task["outputs"]["model"]["permissions"]
-    elif task["category"] == ComputeTaskRep.Category.TASK_AGGREGATE:
-        perm = task["outputs"]["model"]["permissions"]
-    elif task["category"] == ComputeTaskRep.Category.TASK_COMPOSITE:
-        perm = task["outputs"]["shared"]["permissions"]
-    else:
-        raise Exception(f"Unknown parent task category: {task['category']}")
-
-    if from_orchestrator:
-        return perm["process"]
-    else:
-        # for results coming from the cache, there's no distinction between process and download permissions
-        return perm
 
 
 def _compute_parent_task_keys(orc_task, task_data, batch):
@@ -234,6 +139,7 @@ def _register_in_orchestrator(tasks_data, channel_name):
             "category": task_data["category"],
             "algo_key": task_data.get("algo_key"),
             "compute_plan_key": task_data["compute_plan_key"],
+            "outputs": task_data.get("outputs", {}),
             "metadata": task_data.get("metadata") or {},
         }
 
@@ -246,8 +152,6 @@ def _register_in_orchestrator(tasks_data, channel_name):
 
         extra_data_field = EXTRA_DATA_FIELD[orc_task["category"]]
         orc_task[extra_data_field] = _compute_extra_data(orc_task, task_data)
-        trunk_permissions = task_data.get("out_trunk_model_permissions", {})
-        orc_task["outputs"] = _get_task_outputs(channel_name, orc_task, trunk_permissions, batch)
 
         batch[orc_task["key"]] = orc_task